Add settings page and browser action
Add settings page where a user can change the key shortcuts for marking
a location on a page and scroll to it.

Add a browser action for the user for marking a location on a page and
scroll to it using a GUI and to show a captured image of the tab on that
location.